Split up the GroupDetailsViewController into multiple files
